refactor: import services from index file
