+++ src/__tests__/integration/server.test.ts
@@ -7,12 +7,12 @@ import { MemberRepository } from "@/repositories/organizations/members.js";
 import { OrganizationRepository } from "@/repositories/organizations/organizations.js";
 import { UserRepository } from "@/repositories/users/index.js";
 import { feideClient } from "@/services/auth/clients.js";
+import { AuthService } from "@/services/auth/index.js";
 import { FeideProvider } from "@/services/auth/providers.js";
-import { AuthService } from "@/services/auth/service.js";
-import { CabinService } from "@/services/cabins/service.js";
+import { CabinService } from "@/services/cabins/index.js";
 import { MailService } from "@/services/mail/index.js";
-import { OrganizationService } from "@/services/organizations/service.js";
-import { UserService } from "@/services/users/service.js";
+import { OrganizationService } from "@/services/organizations/index.js";
+import { UserService } from "@/services/users/index.js";
 
 import { initServer } from "../../server.js";
