+++ src/index.ts
@@ -8,12 +8,12 @@ import { OrganizationRepository } from "./repositories/organizations/organizatio
 import { UserRepository } from "./repositories/users/index.js";
 import { initServer } from "./server.js";
 import { feideClient } from "./services/auth/clients.js";
+import { AuthService } from "./services/auth/index.js";
 import { FeideProvider } from "./services/auth/providers.js";
-import { AuthService } from "./services/auth/service.js";
-import { CabinService } from "./services/cabins/service.js";
+import { CabinService } from "./services/cabins/index.js";
 import { MailService } from "./services/mail/index.js";
-import { OrganizationService } from "./services/organizations/service.js";
-import { UserService } from "./services/users/service.js";
+import { OrganizationService } from "./services/organizations/index.js";
+import { UserService } from "./services/users/index.js";
 
 const cabinRepository = new CabinRepository(prisma);
 const userRepository = new UserRepository(prisma);
